refactor(static site) allow cross-origin OPTIONS requests to set Access-Control-Allow-Origin header
To support cross-origin ajax requests with credentials, we have to allow the api to
return an Access-Control-Allow-Origin header with the requesting domain set. The more
permissive * value for this header won't work. It's a violation of CORS policy to make a
cross-origin request with credentials with the Access-Control-Allow-Origin set to the *
wildcard.